+++ test/once_connection_once_common_structs/models/common.rs
@@ -13,4 +13,5 @@ pub struct PaginationResult<T> {
     /// Number of pages in total
     pub num_pages: i64,
 }
+
 type Connection = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;