+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This Changelog only lists the changes done to this fork and since 0.0.13
 ## next
 
 - fix that a file is created if the file does not exist on `MarkedFile::new`
+- fix so that all entries in `common.rs` are properly new-lined
 - add option `read-only-prefix` to set prefixes for tables to treat as read-only (like `view_`)
 - add option `no-impls` to not generate any `impl Struct`s (like `read`, `create`, `paginate`)
 - add option `once-connection` to only output the connection type once in `common.rs` instead of in all files